feat: #317 PR에 따른 충돌 해소
