+++ backend/src/room/dto/index.ts
@@ -4,4 +4,4 @@ export * from "./join-room.dto";
 export * from "./move-index.dto";
 export * from "./reaction.dto";
 export * from "./room-id.dto";
-export * from "./room-list.dto";
+export * from "./all-room.dto";
