--- backend/src/room/dto/room-list.dto.ts
@@ -1,14 +1,0 @@
-import { Connection, RoomStatus } from "@/room/domain/room";
-
-export class RoomListResponseDto {
-    createdAt: number;
-    host: Connection;
-    maxParticipants: number;
-    status: RoomStatus;
-    title: string;
-    id: string;
-    category: string[];
-    inProgress: boolean;
-    questionListTitle: string;
-    participants: number;
-}
